Cancel make on live & install wget & ask if user uses uefi and swap

+++ before/cfdisk.c
@@ -2,6 +2,7 @@
 #include <stdio.h>
 #include <string.h>
 char disk1, mnt, esp, swap;
+int ifuefi, ifswap;
 void mountmnt();
 void mountesp();
 void swapon();
@@ -15,6 +16,8 @@ void cfdisk ()
    system("cfdisk ${disk1}");
    system("fdisk -l");
    mountmnt();
+   mountesp();
+   swapon();
 }
 
 void mountmnt ()
@@ -24,26 +27,34 @@ void mountmnt ()
    setenv("mnt", &mnt, 1);
    system("mkfs.ext4 ${mnt}");
    system("mount ${mnt} /mnt");
-   mountesp();
 }
 
 void mountesp ()
 {
-   printf("Type a partition to mount /mnt/boot/efi\n");
-   scanf("%s",&esp);
-   setenv("esp", &esp, 1);
-   system("mkdir /mnt/boot");
-   system("mkdir /mnt/boot/efi");
-   system("mkfs.fat -F32 ${esp}");
-   system("mount ${esp} /mnt/boot/efi");
-   swapon();
+   printf("Do you use UEFI\n Type 1 for yes or 2 for false \n");
+   scanf("%d", &ifuefi);
+   if ( ifuefi == 1 )
+   {
+      printf("Type a partition to mount /mnt/boot/efi\n");
+      scanf("%s",&esp);
+      setenv("esp", &esp, 1);
+      system("mkdir /mnt/boot");
+      system("mkdir /mnt/boot/efi");
+      system("mkfs.fat -F32 ${esp}");
+      system("mount ${esp} /mnt/boot/efi");
+   }
 }
 
 void swapon ()
 {
-   printf("Type a partition to turn on swap\n");
-   scanf("%s",&swap);
-   setenv("swap", &swap, 1);
-   system("mkswap ${swap}");
-   system("swapon ${swap}");
+   printf("Do you use swap\n Type 1 for yes or 2 for false \n");
+   scanf("%d", &ifswap);
+   if ( ifswap == 1 )
+   {
+      printf("Type a partition to turn on swap\n");
+      scanf("%s",&swap);
+      setenv("swap", &swap, 1);
+      system("mkswap ${swap}");
+      system("swapon ${swap}");
+   }      
 }
